+++ eyes_selenium/applitools/common/ultrafastgrid/config.py
@@ -74,9 +74,36 @@ class DeviceName(Enum):
     Pixel_5 = "Pixel 5"
 
 
+class AndroidVersion(Enum):
+    LATEST = "latest"
+    ONE_VERSION_BACK = "latest-1"
+    TWO_VERSIONS_BACK = "latest-2"
+
+
 class IosVersion(Enum):
     LATEST = "latest"
     ONE_VERSION_BACK = "latest-1"
+
+
+class AndroidDeviceName(Enum):
+    Pixel_3_XL = "Pixel 3 XL"
+    Pixel_4 = "Pixel 4"
+    Pixel_4_XL = "Pixel 4 XL"
+    Galaxy_Note_8 = "Galaxy Note 8"
+    Galaxy_Note_9 = "Galaxy Note 9"
+    Galaxy_S8 = "Galaxy S8"
+    Galaxy_S8_Plus = "Galaxy S8 Plus"
+    Galaxy_S9 = "Galaxy S9"
+    Galaxy_S9_Plus = "Galaxy S9 Plus"
+    Galaxy_S10 = "Galaxy S10"
+    Galaxy_S10_Plus = "Galaxy S10 Plus"
+    Galaxy_Note_10 = "Galaxy Note 10"
+    Galaxy_Note_10_Plus = "Galaxy Note 10 Plus"
+    Galaxy_S20 = "Galaxy S20"
+    Galaxy_S20_PLUS = "Galaxy S20 Plus"
+    Galaxy_S21 = "Galaxy S21"
+    Galaxy_S21_PLUS = "Galaxy S21 Plus"
+    Galaxy_S21_ULTRA = "Galaxy S21 Ultra"
 
 
 class IosDeviceName(Enum):